Add opt-in --skip-validators flag to skip function validator compilation
By default every confect query/mutation/action compiles its args/returns
Schema into Convex validators at registration, which runs on every cold
start. Since confect already decodes/encodes via Effect Schema in each
handler, those Convex validators are redundant for confect's own
correctness.

Add a build-time `--skip-validators` flag (with a CONFECT_SKIP_VALIDATORS
env fallback) to `confect codegen`/`confect dev`. When set, codegen emits
registries that route through new validator-free builders, so the
args/returns validator compilation is tree-shaken out and never runs at
registration, trimming per-function cold-start time.

- Split the validator-free handler core out of RegisteredConvexFunction /
  RegisteredNodeFunction (new *Core modules with no SchemaToValidator
  import); the compiling builders inject the Schema->Validator compilers
  via a new RegisteredFunction.Compilers seam, the *WithoutValidators
  builders omit them.
- Thread the flag through codegen via a SkipValidators Context.Reference
  (default off) into the registeredFunctionsForGroup template.
- Tables and the deploy schema are unchanged; Convex still validates
  stored documents.
