+++ packages/server/src/node.ts
@@ -1,1 +1,2 @@
 export * as RegisteredNodeFunction from "./RegisteredNodeFunction";
+export * as RegisteredNodeFunctionWithoutValidators from "./RegisteredNodeFunctionWithoutValidators";
